# product-microservice/src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {
						"type": "boolean",
						"enum": [
							"true"
						],
						"nullable": false
					}
				},
				"required": [
					"message",
					"error"
				],
				"type": "object"
			},
			"Products.ViewModel": {
				"properties": {
					"image": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"shortDescription": {
						"type": "string"
					}
				},
				"required": [
					"image",
					"price",
					"quantity",
					"description",
					"shortDescription"
				],
				"type": "object"
			},
			"Product": {
				"properties": {
					"image": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"shortDescription": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"image",
					"price",
					"description",
					"shortDescription",
					"quantity"
				],
				"type": "object"
			},
			"ProductDto": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Product"
					},
					{
						"properties": {
							"status": {
								"type": "string",
								"enum": [
									"Started",
									"Sent",
									"Derived",
									"Returned"
								]
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"status"
						],
						"type": "object"
					}
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ms_node_1",
		"version": "0.1.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/product": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Products.ViewModel"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"items": {
												"$ref": "#/components/schemas/ProductDto"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/product/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/ProductDto"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/product/{id}/add": {
			"post": {
				"operationId": "Add",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "quantity",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product/{id}/buy": {
			"post": {
				"operationId": "Buy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "quantity",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/events/list/{type}/{id}": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"items": {
												"properties": {},
												"type": "object"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"Product"
							]
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}